+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Don’t use the root logger. Instead, use iatikit’s own logger, and default to NullHandler. ([#46](https://github.com/pwyf/iatikit/issues/46))
+
 ## [2.2.2] – 2019-05-16
 
 ### Added